requirements.txt
Django==5.1.14
dj-database-url==2.3.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-pandas==0.6.7
django-ratelimit==4.1.0
django-silk==5.1.0
deepface==0.0.93
imutils==0.5.4
matplotlib==3.8.4
celery==5.4.0
redis==5.2.1
sentry-sdk[django]==2.43.0
seaborn==0.13.2
Pillow==10.3.0
opencv-python==4.9.0.80
scikit-learn==1.5.0
tf-keras==2.20.1
numpy==1.26.4
pandas==2.2.3
scipy==1.14.1
prometheus-client==0.20.0
PyYAML==6.0.2
cryptography==46.0.3
psycopg2-binary==2.9.9
faiss-cpu==1.9.0.post1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

